+++ hutool-core/src/main/java/cn/hutool/core/util/NumberUtil.java
@@ -2240,19 +2240,14 @@ public class NumberUtil {
 			return BigDecimal.ZERO;
 		}
 
-		try {
-			// 支持类似于 1,234.55 格式的数字
-			final Number number = parseNumber(numberStr);
-			if (number instanceof BigDecimal) {
-				return (BigDecimal) number;
-			} else {
-				return new BigDecimal(number.toString());
-			}
-		} catch (Exception ignore) {
+		try{
+			return new BigDecimal(numberStr);
+		} catch (Exception ignore){
 			// 忽略解析错误
 		}
 
-		return new BigDecimal(numberStr);
+		// 支持类似于 1,234.55 格式的数字
+		return toBigDecimal(parseNumber(numberStr));
 	}
 
 	/**