+++ hutool-core/src/main/java/cn/hutool/core/io/FileUtil.java
@@ -1607,6 +1607,11 @@ public class FileUtil extends PathUtil {
 			return null;
 		}
 
+		//兼容Windows下的共享目录路径（原始路径如果以\\开头，则保留这种路径）
+		if (path.startsWith("\\\\")) {
+			return path;
+		}
+
 		// 兼容Spring风格的ClassPath路径，去除前缀，不区分大小写
 		String pathToUse = StrUtil.removePrefixIgnoreCase(path, URLUtil.CLASSPATH_URL_PREFIX);
 		// 去除file:前缀
@@ -1621,10 +1626,6 @@ public class FileUtil extends PathUtil {
 		pathToUse = pathToUse.replaceAll("[/\\\\]+", StrUtil.SLASH);
 		// 去除开头空白符，末尾空白符合法，不去除
 		pathToUse = StrUtil.trimStart(pathToUse);
-		//兼容Windows下的共享目录路径（原始路径如果以\\开头，则保留这种路径）
-		if (path.startsWith("\\\\")) {
-			pathToUse = "\\" + pathToUse;
-		}
 
 		String prefix = StrUtil.EMPTY;
 		int prefixIndex = pathToUse.indexOf(StrUtil.COLON);