+++ hutool-core/src/test/java/cn/hutool/core/io/FileUtilTest.java
@@ -182,7 +182,9 @@ public class FileUtilTest {
 		Assert.assertEquals("../../bar", FileUtil.normalize("../../bar"));
 		Assert.assertEquals("C:/bar", FileUtil.normalize("/C:/bar"));
 		Assert.assertEquals("C:", FileUtil.normalize("C:"));
-		Assert.assertEquals("\\/192.168.1.1/Share/", FileUtil.normalize("\\\\192.168.1.1\\Share\\"));
+
+		// issue#3253，smb保留格式
+		Assert.assertEquals("\\\\192.168.1.1\\Share\\", FileUtil.normalize("\\\\192.168.1.1\\Share\\"));
 	}
 
 	@Test