+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * 【core  】      ZipReader增加setMaxSizeDiff方法，自定义或关闭ZipBomb（issue#3018@Github）
 * 【db    】      Query.of(entity)构建时传入fields（issue#I7M5JU@Gitee）
 * 【db    】      clickhouse驱动名称变更为com.clickhouse.jdbc.ClickHouseDriver（issue#3224@Github）
+* 【core  】      UrlResource增加size方法（issue#3226@Github）
 
 ### 🐞Bug修复
 * 【core  】      修复MapUtil工具使用filter方法构造传入参数结果问题（issue#3162@Github）